+++ test/module/resource-tracker/migration.test.ts
@@ -1,4 +1,4 @@
-import { migrateTemplateToV2, migrateTrackerInstanceToV2 } from '@module/resource-tracker/migrations/1_0_0.js'
+import { migrateTemplateToV2, migrateTrackerInstanceToV2 } from '@module/resource-tracker/migrations/v1_0_0.js'
 import { IResourceTracker } from '@module/resource-tracker/types.js'
 import { vi } from 'vitest'
